refactor: rename PhysicalTenantIdDiscovery to PhysicalTenantConfigUtil
The class no longer only discovers tenant ids — since the walk primitive was
extracted it also iterates the properties each tenant declares. "IdDiscovery"
undersold that; PhysicalTenantConfigUtil names it for what it is: a stateless
utility over physical-tenant configuration (id discovery, property iteration,
and id validation). Pure rename; no behavior change.
