+++ authentication/src/test/java/io/camunda/authentication/pt/PhysicalTenantAuthConfigurationsTest.java
@@ -7,7 +7,7 @@
  */
 package io.camunda.authentication.pt;
 
-import static io.camunda.spring.utils.PhysicalTenantIdDiscovery.MAX_TENANT_ID_LENGTH;
+import static io.camunda.spring.utils.PhysicalTenantConfigUtil.MAX_TENANT_ID_LENGTH;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
